+++ package.json
@@ -2,7 +2,7 @@
   "name": "soundboard",
   "version": "1.0.0",
   "description": "final attempt",
-  "main": "server.js",
+  "main": "app.js",
   "scripts": {
     "test": "mocha",
     "start": "node server.js"
@@ -11,60 +11,23 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "babel-cli": "6.16.0",
-    "babel-core": "6.17.0",
-    "babel-loader": "6.2.5",
-    "babel-preset-latest": "6.16.0",
-    "babel-register": "6.16.3",
-    "chai": "^3.5.0",
-    "chalk": "1.1.3",
-    "cheerio": "0.22.0",
-    "compression": "1.6.2",
-    "cross-env": "3.1.3",
-    "css-loader": "0.25.0",
-    "eslint": "3.8.1",
-    "eslint-plugin-import": "2.0.1",
-    "eslint-watch": "2.1.14",
-    "express": "^4.16.2",
-    "extract-text-webpack-plugin": "1.0.1",
-    "html-webpack-plugin": "2.22.0",
-    "jsdom": "9.8.0",
-    "json-schema-faker": "0.3.6",
-    "json-server": "0.8.22",
-    "localtunnel": "1.8.1",
-    "mocha": "^3.5.3",
-    "nock": "8.1.0",
-    "npm-run-all": "3.1.1",
-    "nsp": "2.6.2",
-    "numeral": "1.5.3",
-    "open": "0.0.5",
-    "rimraf": "2.5.4",
-    "style-loader": "0.13.1",
-    "webpack": "1.13.2",
-    "webpack-dev-middleware": "1.8.4",
-    "webpack-hot-middleware": "2.13.0",
-    "webpack-md5-hash": "0.0.5"
+    "chai": "^4.1.2",
+    "chai-http": "^3.0.0",
+    "mocha": "0.0.1"
   },
   "directories": {
     "test": "test"
   },
   "dependencies": {
-    "bcryptjs": "^2.4.3",
     "body-parser": "^1.18.2",
     "bootstrap": "^3.3.7",
-    "chai-http": "^3.0.0",
-    "connect": "^2.30.2",
-    "connect-rest": "^3.0.13",
+    "connect-rest": "^3.0.12",
     "cors": "^2.8.4",
-    "elasticsearch": "^12.1.3",
-    "jsonwebtoken": "^8.2.0",
-    "mongoose": "^5.0.9",
-    "mongoose-elasticsearch-xp": "^5.4.1",
+    "express": "^4.16.2",
+    "mongoose": "^5.0.5",
     "morgan": "^1.9.0",
-    "nodemailer": "^4.6.2",
-    "passport": "^0.4.0",
+    "nodemailer": "^4.4.2",
     "request": "^2.83.0",
-    "server.js": "^1.0.0",
     "session-mongoose": "^0.5.2",
     "uuid": "^3.2.1"
   },
