+++ public/user:index.js
@@ -1,150 +1,113 @@
-const serverBase = '//localhost:8080:';
-const USER_URL = "./server.js";
-const EVENT_URL = serverBase + "./eventsRouter.js";
-const ARTIST_URL = serverBase + "./eventsRouter.js"
-
 function handleNavigation() {
-    $('nav a').click(function(event) {
+
+    $("nav a").click(function(event) {
         event.preventDefault();
-        const currentPage = $(this).data('page');
-    $('.page').hide();
+        const currentPage = $(this).data("page");
+    $(".page").hide();
     $(`.${currentPage}`).show();
   });
-  }
 
     function openNav() {
-        document.getElementById("nav_start").style.width = "15%";
+        document.getElementById("mySidenav").style.width = "15%";
         document.getElementById("main").style.marginLeft = "15%";
     }
     // Clear form values
 
+    
     function setEventListeners() {
         handleNavigation();
     }
 
     //show start page
     function showSignUpPage() {
-        $(".signUpPage").show();
         $(".page").hide();
-  
+        $(".sign_up_page").show();
     }
-
-    function toggleRegistration() {
-      $('#signInLink').show();
-      $('.link').click(function(event) {
-        event.preventDefault();
-        $('#signInLink').hide();
-        $('#signInPage').show();
-        ${'#signUpPage'}.hide();
-      })
-    };
-
-    const userTemplate = (
-  '<div class="user_js">' +
-    '<h3 class="user_name"><h3>' +
-    '<hr>' +
-    '<ul class="user_info">' +
-    '</ul>' +
-    '</div>' +
-  '</div>'
-);
-
-     
-function handleNewUser() {
- 
-    const usernew = $('input[name=username]');
-
-   
-
-      const passnew = $('input[name=password]');
-       
     
- 
-     const emailnew = $('input[name=email]');
-     storeUser();
-     }
-  }
-}
-  function storeUser() {
-    const user = {
-      username: usernew,
-      password: passnew,
-      email: emailnew
-      };
-      return user;
-      postNewUser();
-     };
-   
-    function postNewUser() {
-       $.ajax({
-        method: 'POST',
-        url: 'https://mongodb://mbarker1221:shompin1@ds131698.mlab.com:31698/users';
-        req.params.body = {
-           username: usernew,
-           password: passnew,
-           email: emailnew
-        }
-        dataType: 'json',
-        contentType: 'application/json'
-        success: return user;
-        renderSearchPage();
-      }),
-  };
+    function validateUsername() {}
 
-function handleOldUser() {
+    
+    function validatePass() {}
 
-    function renderSignInPage() {
-      const userN=('input[name=un');
-      if (userN.length < 5) {
-      alert('must be more than 4 characters');
-      }
-       if (userN.length > 15) {
-        alert('must be less than 16 characters')
-       }
-       if (userN.length > 4 && < 16){
-      const userName= userN
-    };
-
-      const passW=('imput[name=pw');
-       if (passW.length < 5) {
-        alert('must be more than 4 characters');
-       }
-       if (passW.length > 15) {
-        alert('must be less than 16 characters')
-       }
-       if (passW.length > 4 && < 16) {
-        const passWord= passW
-       } 
-        authUser()
-    };
-
-      function authUser() {
-         $.ajax({
-        method: 'POST',
-        url: USER_URL,
-        req.params.body = {
-           username: userName,
-           password: passWord,
-        }
-        dataType: 'json',
-        contentType: 'application/json'
-        success: return user;
-        renderSearchPage();  
-        })
-      };
-
+    
+    function validateEmail() {}
         
-    function renderSearchPage() {
-     
-      }
+
+    function authorizeUsername() {}
+
+    
+    function authorizePass() {}
+      
+
+    function createUser() {
+       const userUN = $('input[name=username]');
+       const userPASS = $('input[name=password]');
+       const userEMAIL = $('input[name=email]');
+
+       const userNew = {
+          username: userUN,
+          password: userPASS,
+          email: userEMAIl
+       }
+
+         $.ajax({
+            type: "POST",
+            url: "/server.js",
+            query: ${userNew},
+            timeout: 2000,
+
+          success: function() {
+              alert('Success!')
+          },
+
+          error: function() {
+              alert('Error!')
+          }
+      });
     }
 
+    function findUser() {
+      const yourUsername = $('input[name=un]');
+      const yourPassword = $('input[name=pass]');
+      const userId = {
+        username: yourUsername,
+        password: yourPassword
+      }
+
+      $.ajax({
+          type: "GET",
+          url: "/server.js",
+          query: ${userId},
+          timeout: 2000,
+
+        success: function() {
+            alert('Success!')
+        },
+        error: function() {
+            alert('Error!')
+        }
+      });
+    }
+
+    //retrieve data from client side
   function getArtist() {
     const artistIn = $('input[name=artistSearch]');
-        $.getJSON(ARTIST_URL, artistIn, function(data) {
-          showArtist(data); 
-        })
+        
+      $.ajax({
+        type: "POST",
+        url: "/eventsRouter.js",
+        query: ${artistIn},
+        timeout: 2000,
+
+      success: function() {
+        alert('Success!')
+      },
+
+      error: function() {
+        alert('Error!')
       }
+    });
+  }
     
     //display artist results page 
   function showArtist(results) {
@@ -161,12 +124,24 @@ function handleOldUser() {
       //retrieve data from client side 
   function getEvents() {
     const eventIn = $('input[name=eventSearch]');
-    const eventLoc = eventIn.val();
-    $.getJSON(EVENT_URL, eventLoc, function(data) {
-      showEvents(data);
-    })
-  }
-   
+         
+         $.ajax({
+            type: "POST",
+            url: "/EventsRouter.js",
+            query: ${eventIn},
+            timeout: 2000,
+
+          success: function() {
+            //show content
+            alert('Success!')
+          },
+
+          error: function() {
+            //show error message
+            alert('Error!')
+          }
+        });
+    }
 
      //display location results page
     function showEvents(results) {
@@ -181,47 +156,6 @@ function handleOldUser() {
       $(".locShows").html(loc);
     };
 
-  
-
-  function findUser() {
-      const yourUsername = $('input[name=un]');
-      const yourPassword = $('input[name=pass]');
-      const userCred = {
-        username: yourUsername,
-        password: yourPassword
-      }
-
-      $.ajax({
-          type: "POST",
-          url: "USER_URL",
-          query: ${userCred},
-
-        success: function() {
-            alert('Success!')
-        },
-        error: function() {
-            alert('Error!')
-        }
-      });
-    }
-
-         
-        /* $.ajax({
-            type: " POST",
-            url: "EVENT_URL",
-            query: eventLoc,
-
-          success: function() {
-            alert('Success!')
-            showEvents(data);
-          },
-
-          error: function() {
-            alert('Error!')
-          }
-        });
-    }*/
-
       //profile page  
     function getUser() {
       const currUN = $('input[name=currentUN]');
@@ -232,17 +166,21 @@ function handleOldUser() {
           username: currUN,
           password: currPASS,
           email: currEMAIL
-        };
+        }
 
     $.ajax({
-        type: " POST",
-        url: "USER_URL",
+        type: "POST",
+        url: "/server.js",
         query: ${you},
+        timeout: 2000,
 
       success: function() {
+          //show content
          alert('Success!')
       },
+
       error: function() {
+          //show error message
          alert('Error!')
       }
     });
@@ -259,46 +197,90 @@ function handleOldUser() {
         const newUsername = $('input[name=editUN]');
         const newPass = $('input[name=editPASS]');
         const newEmail = $('input[name=editEMAIL]');
+
         const newYOU = {
           username: newUsername,
           password: newPass,
           email: newEmail
-        };
+        }
 
          $.ajax({
             type: "POST",
-            url: "USER_URL",
+            url: "/server.js",
             query: ${newYOU},
+            timeout: 2000,
 
             success: function() {
+              //show content
               alert('Success!')
             },
 
             error: function() {
+              //show error message
               alert('Error!')
             }
-        });
-      
+        }
+      });
     }
-    function deleteUser() {
-      const userId = user.id 
-    };
 
-        $.ajax({
+    //postItem()
+     app.put('/user', function(req,res) {
+        db.users.this(user),
+          function(err, users) {
+            let context = {
+              user: user.map(function(user) {
+              return{
+                username: this.username,
+                password: this.password,
+                email: this.email,} 
+              })
+            }
+           }
+
+        if (!(req.params.id && req.body.id && req.params.id === req.body.id)) {
+          const message = (`information is not a match`);
+            console.error(message);
+            return res.status(400).json({
+              message: message
+            });
+          }
+            const toUpdate = {};
+            const updateableFields = ['username', 'password', 'email'];
+
+            updateableFields.forEach(field => {
+              if (field in req.body) {
+                toUpdate[field] = req.body[field]
+              }
+            });
+
+        User
+          .this(req.params.id, {$set: toUpdate})
+          .then(user => res.status(204) 
+          .catch(err => res.status(500).json({
+              message: 'Internal server error'
+          }))
+        )
+      });
+    };
+     
+     function deleteUser() {
+      const userDel = $('input[name=eventSearch]');
+         $.ajax({
             type: "POST",
-            url: "USER_URL",
-            query: ${user.id},
+            url: "/eventsRouter.js",
+            query: ${eventIn},
+            timeout: 2000,
 
             success: function() {
-              alert('Success!')
+                alert('Success!')
             },
-
             error: function() {
-              alert('Error!')
+                alert('Error!')
             }
         });
-      }
-    
+     }
+
+       // On load event:
       $(() => {
         setEventListeners();
         showSignUpPage();
